feat: add validation to ensure that only the broker of the property or an admin can modify it

+++ src/main/java/com/mlcdev/realestate/service/PropertyService.java
@@ -5,9 +5,11 @@ import com.mlcdev.realestate.dto.PropertyDetailDTO;
 import com.mlcdev.realestate.dto.PropertyPatchDTO;
 import com.mlcdev.realestate.dto.PropertySummaryDTO;
 import com.mlcdev.realestate.entities.Property;
+import com.mlcdev.realestate.exception.BusinessRuleException;
 import com.mlcdev.realestate.exception.NotFoundException;
 import com.mlcdev.realestate.mapper.PropertyMapper;
 import com.mlcdev.realestate.repository.PropertyRepository;
+import com.mlcdev.realestate.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
@@ -24,6 +26,7 @@ public class PropertyService {
 
     private final PropertyRepository propertyRepository;
     private final ImageService imageService;
+    private final UserRepository userRepository;
 
     @Transactional(readOnly = true)
     public Page<PropertySummaryDTO> findAll(Pageable pageable){
@@ -39,34 +42,45 @@ public class PropertyService {
     }
 
     @Transactional
-    public PropertyDetailDTO create(PropertyCreateDTO createDTO){
+    public PropertyDetailDTO create(PropertyCreateDTO createDTO, UUID brokerId){
         log.info("Creating property with title: {}", createDTO.getTitle());
-        Property savedProperty = propertyRepository.saveAndFlush(PropertyMapper.createDTOToEntity(createDTO));
+        Property property = PropertyMapper.createDTOToEntity(createDTO);
+        property.setBroker(userRepository.getReferenceById(brokerId));
+        Property savedProperty = propertyRepository.saveAndFlush(property);
         log.info("Property successfully created with ID: {}", savedProperty.getId());
         return PropertyMapper.entityToDetailDTO(savedProperty);
     }
 
     @Transactional
-    public PropertyDetailDTO update(UUID id, PropertyPatchDTO dto){
-        log.info("Patching property with id: {}", id);
-        Property property = propertyByIdOrElseThrow(id);
+    public PropertyDetailDTO update(UUID propertyId, PropertyPatchDTO dto, UUID brokerId, boolean isAdmin){
+        log.info("Patching property with id: {}", propertyId);
+        Property property = propertyByIdOrElseThrow(propertyId);
+        validateUserAdminOrPropertyBroker(property, brokerId, isAdmin);
         Property updatedProperty = propertyRepository.saveAndFlush(PropertyMapper.applyPatchDTOToEntity(dto, property));
-        log.info("Property with ID: {} successfully patched", id);
+        log.info("Property with ID: {} successfully patched", propertyId);
         return PropertyMapper.entityToDetailDTO(updatedProperty);
     }
 
     @Transactional
-    public void delete(UUID id) {
-        log.info("Deleting property with id: {}", id);
-        Property property = propertyByIdOrElseThrow(id);
-        imageService.deleteAllImagesForProperty(id);
+    public void delete(UUID propertyId, UUID brokerId, boolean isAdmin) {
+        log.info("Deleting property with id: {}", propertyId);
+        Property property = propertyByIdOrElseThrow(propertyId);
+        validateUserAdminOrPropertyBroker(property, brokerId, isAdmin);
+        imageService.deleteAllImagesForProperty(propertyId);
         propertyRepository.delete(property);
-        log.info("Property with ID: {} successfully deleted", id);
+        log.info("Property with ID: {} successfully deleted", propertyId);
     }
 
     private Property propertyByIdOrElseThrow(UUID id){
         return propertyRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException("Property with ID: " + id + " not found"));
+    }
+
+    private void validateUserAdminOrPropertyBroker(Property property, UUID brokerId, boolean isAdmin){
+        if(!isAdmin && !property.getBroker().getId().equals(brokerId)){
+            log.warn("User with ID: {} doesn't have the permission to modify the property with ID: {}", brokerId, property.getId());
+            throw new BusinessRuleException("User doesn't have the permission to modify the property");
+        }
     }
 }
